Fix the lifetime of `progam_iter` items
The `ProgramHeader` objects produced by `ElfFile::program_iter` should
have the same lifetime as the `ElfFile`, i.e. the lifetime of the input
data. Prior to this commit, they instead had the shorter of a) the
`ElfFile` lifetime and b) the `&self` reference lifetime, making code
dealing with these `ProgramHeader`s awkward to write.

This commit fixes the lifetime issue by making the lifetime annotations
the same as for `ElfFile::section_iter`.